+++ src/include/script_storge/character_sprite.h
@@ -10,7 +10,7 @@ class Character : public Sprite{
         Properties_Character* derived_properties;
         bool is_prime_sprite = false;
     public:
-        Character(long ID, bool is_prime = true, Properties_Character* properties_ptr = nullptr, SpriteType sprite_type = SpriteType::Character);
+        Character(uint16_t ID, bool is_prime = true, Properties_Character* properties_ptr = nullptr, SpriteType sprite_type = SpriteType::Character);
 
         const Properties_Character* getProperties() override;
         void whenMovedSprite(int x, int y);
