+++ src/include/script_storge/player.h
@@ -9,7 +9,7 @@ class Player : public Character{
         Properties_Player* derived_properties;
         bool is_prime_sprite = false;
     public:
-        Player(long ID, bool is_prime = true, Properties_Player* properties_ptr = nullptr, SpriteType sprite_type = SpriteType::Player);
+        Player(uint16_t ID, bool is_prime = true, Properties_Player* properties_ptr = nullptr, SpriteType sprite_type = SpriteType::Player);
 
         const Properties_Player* getProperties() override;
         ~Player();
